+++ LearnComputer/LearnComputer/CircuitInfrustructure/Basic/InputEndpoint.cs
@@ -4,10 +4,10 @@ namespace LearnComputer.CircuitInfrustructure
 {
     public class InputEndpoint: Endpoint, IInputEndpoint
     {
-        public delegate void ReceiveSignalHanlder(Endpoint sender, Byte signal);
+        public delegate void ReceiveSignalHanlder(Endpoint sender, Int32 signal);
 
         public event ReceiveSignalHanlder Receive = (sender, signal) => { };
-        public Byte LastReceivedSignal { get; private set; }
+        public Int32 LastReceivedSignal { get; private set; }
 
         public InputEndpoint():base(null)
         {}
@@ -15,7 +15,7 @@ namespace LearnComputer.CircuitInfrustructure
         public InputEndpoint(Endpoint connectTo):base(connectTo)
         {}
 
-        public void Transmit(Byte signal)
+        public void Transmit(Int32 signal)
         {
             if (signal != 0 && signal != 1)
                 throw new ArgumentException(String.Format(INVALID_SIGNAL_EXCEPTION_FORMAT, signal));
